Move entity related elements into DashboardDetailsEntityView component.
